update readme. Code clean up
